LayoutTests/media/track/track-in-band-cues-added-once-expected.txt

In-band text tracks.

EVENT(addtrack)
EXPECTED (event.track == 'video.textTracks[0]') OK

EVENT(canplaythrough)

** Check initial in-band track states
EXPECTED (video.textTracks.length == '1') OK
RUN(inbandTrack1 = video.textTracks[0])
EXPECTED (inbandTrack1.mode == 'disabled') OK
EXPECTED (inbandTrack1.cues == 'null') OK
EXPECTED (inbandTrack1.language == 'en') OK
EXPECTED (inbandTrack1.kind == 'captions') OK

RUN(inbandTrack1.mode = 'showing')
RUN(video.play())
EXPECTED (inbandTrack1.cues.length > '0') OK
RUN(video.pause())
RUN(video.currentTime = 0)
RUN(video.play())

EVENT(seeked)
EXPECTED (inbandTrack1.cues.length > '0') OK
RUN(video.pause())
END OF TEST
